Add local login strategy
Our trusted React Client needs a way to authenticate to the API but it
cannot store client secrets. We can store our API client information
on the server of the API, and allow the React client to authenticate
with a username and password. Thsi route is mostly a facade that will
delegate the logic the an OAuth2 password grant flow, but is more secure
because the client information is stored on the server.