Serialize/Deseerialize redux store to local storage
It is necessary to have the ability to persist some of the state of the
redux store for page refreshes. Without client-side persistence of
select data there would be no simple way for a user to stay logged in if
they were to refresh their page. This is just one example use-case but
it is obvious that this is something that will be needed. There is also
a throttling mechanism because the serialization to localstorage is
expensive, so, at most, this can write to local storage once every
second. This will allow us to capture most of the data, but still stay
performant.