Refactor controllers to contain their routes
The way the routing was originally occurring is not going to scale well
now that there are a few more controllers. Now each controller should
export a function that takes the router as a parameter and the
controller should declare all of the routes within this function. All
request handlers should not be exported from the controller file.